+++ packages/dso-toolkit/src/components/form/models/form-group-date-picker-legacy.model.ts
@@ -1,0 +1,9 @@
+import { DatePickerLegacy } from "../../date-picker-legacy/date-picker-legacy.models.js";
+
+import { FormGroupBase } from "./form-group.base-model.js";
+
+export interface FormGroupDatePickerLegacy<TemplateFnReturnType>
+  extends Omit<FormGroupBase<TemplateFnReturnType>, "label"> {
+  group: "date-picker-legacy";
+  datePickerLegacy: DatePickerLegacy;
+}